Change TextDictionary Population + Read char desc from there (#19)
* Change MPQ text reading

Previously checked for there to be exactly 3 commas, changed to check that line doesn't start with hash or slash (includes/comments) + that it has 3 or more commas.

This might cause issues later, but it's needed for now otherwise lines get skipped.

* Read in Not Xpac Char desc from dictionary

Also added a string utils class. How do I test with .Net?