d2input: MIDI as an example input backend
Changing the value of key "Input" in config.json to 'midi' will use portmidi as
the input backend. Currently, it just attaches to the default MIDI output of the
system. MIDI patchbay software exists for Mac, Windows, and Linux which can be
used for routing specific midi hardware/software into the default system midi
ports.

MIDI Note On/Off messages will press keys. Notes 1 thru 125 will send whatever
characters correspond to those keycode integers. Mouse buttons are the notes
126 (Left), 127 (Middle), and 128 (Right). Mouse cursor x-axis is MIDI-CC 7,
and y-axis is CC 8.

part of this commit touches d2ui and d2gamescene because of the references
to ebiten which were removed, and are now handled by the input manager events.

Caveats:
* This has only been tested on debian linux, not sure if it works anywhere else
* Character input to the console is not implemented by this backend
* Can't use system keyboard if MIDI is specified as the input backend