AppComponent interface
Added an `AppComponent` interface for singletons to implement. The goal is to
try and make these `AppComponent` implementation modular, but with the ability
to call upon functionality of other `AppComponent` instances without exploding.

The `App` initializes `AppComponents` and calls the Advance/Render methods in
its update loop. The order in which they are processed is currently determined
by the name key of the component, which is just a string.

The commit is large because I was initially trying to implement this with
d2input, which is touched by a lot of other files. Since I defined interfaces
for input stuff, I had to go update all of those interface references.