AppComponents implementation
Had to implement this to sort out the initialization ordering of
audio/input/render/terminal/gui etc. Basically, the main app creates
instances of the various components, then a two-way reference is created
between the app and the components. Before any of the components start
getting used, they get initialized. Because they have a reference to the app,
they now have a means of retrieving other AppComponents.